# back/api/openapi.yaml
openapi: 3.0.3
info:
  title: Plant Management API
  description: API REST pour la gestion de plantes domestiques
  version: 1.0.0
  contact:
    name: Plant Team
servers:
  - url: http://localhost:8080/api/v1
    description: Development (Quarkus local)
  # a remplacer par l'URL du cluster Kubernetes université ?
  - url: http://plant-backend.etu-k8s.lan/api/v1
    description: Production (Kubernetes cluster)
...
security:
  - bearerAuth: []
paths:
...
  /plants/{id}:
    get:
      tags: [Plant]
      summary: Détails d'une plante
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Détails de la plante
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlantDetailResponse'
    patch:
      tags: [Plant]
      summary: Modifier une plante
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePlantRequest'
      responses:
        '200':
          description: Plante modifiée
    delete:
      tags: [Plant]
      summary: Supprimer une plante
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Plante supprimée
...
  /notifications:
    get:
      tags: [Notification]
      summary: Mes notifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Liste des notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationResponse'
  /notifications/{id}/read:
    post:
      tags: [Notification]
      summary: Marquer comme lue
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Notification marquée comme lue
...
  /mcp/tools:
    post:
      tags: [MCP]
      summary: Exécuter un outil MCP
      description: |
        Endpoint pour l'intégration LLM via Goose/Mistral.
        Permet de créer/lister des plantes via commandes naturelles.
      security:
        - mcpApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/McpToolRequest'
            examples:
              search:
                summary: Rechercher des plantes
                value:
                  tool: "search_plants"
                  params:
                    query: "rose"
              add:
                summary: Ajouter une plante
                value:
                  tool: "add_plant"
                  params:
                    speciesName: "Rose"
                    roomName: "Salon"
                    nickname: "Belle"
              list:
                summary: Lister les plantes
                value:
                  tool: "list_plants"
                  params: {}
              water:
                summary: Arroser une plante
                value:
                  tool: "water_plant"
                  params:
                    plantName: "Belle"
      responses:
        '200':
          description: Outil exécuté avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/McpToolResponse'
        '400':
          description: Outil inconnu ou paramètres invalides
        '401':
          description: API Key MCP invalide
        '404':
          description: Ressource non trouvée (plante, espèce, etc.)
  /mcp/schema:
    get:
      tags: [MCP]
      summary: Obtenir le schéma des outils MCP
      description: |
        Retourne la liste des outils disponibles avec leurs paramètres.
        Utilisé par Goose pour configurer l'intégration.
      security:
        - mcpApiKey: []
      responses:
        '200':
          description: Schéma des outils MCP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/McpSchema'
        '401':
          description: API Key MCP invalide
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    mcpApiKey:
      type: apiKey
      in: header
      name: X-MCP-API-Key
      description: API Key pour l'accès MCP (Goose/Mistral)
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    RoomResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
        plantCount:
          type: integer
...
    SpeciesResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        trefleId:
          type: integer
        commonName:
          type: string
        scientificName:
          type: string
        family:
          type: string
        imageUrl:
          type: string
...
    UpdatePlantRequest:
      type: object
      properties:
        roomId:
          type: string
          format: uuid
          nullable: true
        nickname:
          type: string
        wateringIntervalDays:
          type: integer
        notes:
          type: string
    PlantResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nickname:
          type: string
        species:
          $ref: '#/components/schemas/SpeciesResponse'
        room:
          $ref: '#/components/schemas/RoomResponse'
        photoUrl:
          type: string
        lastWatered:
          type: string
          format: date-time
        nextWatering:
          type: string
          format: date-time
        needsWater:
          type: boolean
        acquiredAt:
          type: string
          format: date
    PlantDetailResponse:
      allOf:
        - $ref: '#/components/schemas/PlantResponse'
        - type: object
          properties:
            wateringIntervalDays:
              type: integer
            notes:
              type: string
            createdAt:
              type: string
              format: date-time
            recentCareLogs:
              type: array
              items:
                $ref: '#/components/schemas/CareLogResponse'

    # ===== CARE LOG =====
...
    CareLogResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        action:
          type: string
        notes:
          type: string
        performedBy:
          type: object
          properties:
            id:
              type: string
              format: uuid
            displayName:
              type: string
        performedAt:
          type: string
          format: date-time

    # ===== NOTIFICATION =====
    NotificationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [WATERING_REMINDER, CARE_REMINDER, PLANT_ADDED, MEMBER_JOINED]
        message:
          type: string
        read:
          type: boolean
        plantId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time

    # ===== MCP (OBLIGATOIRE) =====
    McpToolRequest:
      type: object
      required: [tool]
      properties:
        tool:
          type: string
          enum: [search_plants, add_plant, list_plants, water_plant]
          description: Nom de l'outil à exécuter
        params:
          type: object
          additionalProperties:
            type: string
          description: Paramètres de l'outil (clé-valeur)
          example:
            speciesName: "Rose"
            roomName: "Salon"
    McpToolResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, error]
        message:
          type: string
          description: Message de résultat lisible
        data:
          type: object
          description: Données retournées (varie selon l'outil)
          additionalProperties: true
    McpSchema:
      type: object
      properties:
        name:
          type: string
          example: "Plant Management MCP"
        version:
          type: string
          example: "1.0.0"
        tools:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              description:
                type: string
              parameters:
                type: object
                additionalProperties:
                  type: object
                  properties:
                    type:
                      type: string
                    required:
                      type: boolean
                    description:
                      type: string